// frontend/src/app/core/api/generated/model/page-response-supplier-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SupplierResponse } from './supplier-response';


export interface PageResponseSupplierResponse { 
    content?: Array<SupplierResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    first?: boolean;
    last?: boolean;
}